src/production_ecommerce/requirements.txt
asgiref==3.2.3
autopep8==1.5.3
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
certifi==2019.11.28
cryptography==2.8
Django
django-bootstrap-modal-forms==1.4.1
django-countries==5.5
django-crispy-forms==1.8.1
django-etc==1.2.0
django-extensions==2.2.6
django-hitcount==1.3.3
django-model-utils==4.0.0
django-resized==0.3.11
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
entrypoints==0.3
fasteners==0.14.1
flake8==3.8.3
future==0.18.2
httplib2==0.14.0
idna==2.8
image==1.5.31
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.6.0
pyflakes==2.2.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-dateutil==2.7.3

pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==3.14.0
simplejson==3.16.0
six==1.13.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.0
toml==0.10.1
